requirements.txt
requests==2.32.3
pandas==2.2.3
pyarrow==18.0.0
duckdb==1.1.3
pyyaml==6.0.2
python-dateutil==2.9.0.post0
tenacity==9.0.0
